Fix broken documentation build (#74)
Release 0.7.1 [failed to build on docs.rs](https://docs.rs/crate/c2pa/0.7.1) because the so-called "no-op" C++ code in XMP Toolkit contained a reference to a field that was disabled elsewhere. (See https://github.com/adobe/xmp-toolkit-rs/pull/67.)

This PR incorporates the fixed version of XMP Toolkit and and updates the PR validation loop to more closely mimic the docs.rs runtime environment so we detect future errors of this type.