C2PA-305: Use of a `enum FontError` in the `font_io` handler. (#31)
Following examples in other areas of the library, this moves the font-related errors to the asset_handlers/font_io.rs module as:

enum FontError.

At this point, keeping the changes small with just the move and keeping the errors the same. This should facilitate a quicker change for the other story which picks up testing in more earnest to be unblocked.

I also noticed the library also has other places with this type of error structure and provides a wrap_*_err fn, so I did the same to try to stay consistent with the other parts of the library.