C2PA-305: More detailed font errors (#33)
An extension to the previous PR which now converts all internal font_io/sft_io/woff_io calls to use FontError as the error type. And then all of the crate level trait implementations, the error is converted to a crate::error::Error.

I also took the time to delete unused errors and split errors that were being used generically into more detailed errors.

For the SaveError I spun out a separate error enum to scope the errors for more details on the save errors.