feat: Simplify `StatusTracker` interface (#937)
Having `StatusTracker` as a non-object-safe trait made it impossible to use in further trait APIs.

See early draft of #936 for an example of this problem.

This PR changes `StatusTracker` to use an enum to describe the only two implementations of `StatusTracker` that have ever existed.

Remove the `take_errors` API in favor of a new `has_any_error` function.