feat!: Add `no_file_io` feature and remove `file_io`
Today the assumption of file system access in the c2pa crate is so
pervasive that tests will fail unless the `file_io` feature is enabled.

This change in feature flags will reflect the reality of this design,
and still allow for use cases in which file I/O is unavailable.

\### Example of cfg statements:
Before
```
    #[cfg(feature = "file_io")]
```
After
```
    feature = "no_file_io",
    all(target_arch = "wasm32", not(target_os = "wasi"))
)))]
```