chore: ensures dylib in release folder is copied (#1188)
The dylib / dlls in the deps file were being copied to the zip
output instead of the articats in the release folder. This was
messing up the paths in the dylibs as we were converting the paths
to relative paths in the dylibs in the release folder, then copying
the dylibs from the deps folder into the zip during release.

Co-authored-by: Dylan Ross <dyross@adobe.com>